🚚 Move: `Prisecter` struct to `pagination` module [#63]
